Featrue/open data (#2771)
feat(h5): open api 功能更新

- api以小程序为准scanCode
- processApis去掉无用分支，增加formatParams 
- 增加h5支付
- 使用最新sdk文档
