+++ docs/native-api.md
@@ -996,7 +996,7 @@ Taro.clearStorageSync()
 
 #### Taro.getLocation(OBJECT)
 
-使用方式同 [`wx.getLocation`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.getLocation.html)，h5端仅支持[微信公众号](https://qydev.weixin.qq.com/wiki/index.php?title=%E5%BE%AE%E4%BF%A1JS%E6%8E%A5%E5%8F%A3#.E8.8E.B7.E5.8F.96.E5.9C.B0.E7.90.86.E4.BD.8D.E7.BD.AE.E6.8E.A5.E5.8F.A3)，支持 `Promise` 化使用。
+使用方式同 [`wx.getLocation`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.getLocation.html)，h5端仅支持[微信公众号]h5端仅支持[微信公众号](https://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421141115)（API以小程序为准），支持 `Promise` 化使用。
 
 **示例代码：**
 
@@ -1029,7 +1029,7 @@ Taro.chooseLocation(params).then(...)
 
 #### Taro.openLocation(OBJECT)
 
-使用方式同 [`wx.openLocation`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.openLocation.html)，h5端仅支持[微信公众号](https://qydev.weixin.qq.com/wiki/index.php?title=%E5%BE%AE%E4%BF%A1JS%E6%8E%A5%E5%8F%A3#.E4.BD.BF.E7.94.A8.E5.BE.AE.E4.BF.A1.E5.86.85.E7.BD.AE.E5.9C.B0.E5.9B.BE.E6.9F.A5.E7.9C.8B.E4.BD.8D.E7.BD.AE.E6.8E.A5.E5.8F.A3)，支持 `Promise` 化使用。
+使用方式同 [`wx.openLocation`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.openLocation.html)，h5端仅支持[微信公众号](https://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421141115)（API以小程序为准），支持 `Promise` 化使用。
 
 **示例代码：**
 
@@ -1385,7 +1385,7 @@ Taro.makePhoneCall(params).then(...)
 
 #### Taro.scanCode(OBJECT)
 
-使用方式同 [`wx.scanCode`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.scanCode.html)，h5端仅支持[微信公众号](https://qydev.weixin.qq.com/wiki/index.php?title=%E5%BE%AE%E4%BF%A1JS%E6%8E%A5%E5%8F%A3#.E8.B0.83.E8.B5.B7.E5.BE.AE.E4.BF.A1.E6.89.AB.E4.B8.80.E6.89.AB.E6.8E.A5.E5.8F.A3)，支持 `Promise` 化使用。
+使用方式同 [`wx.scanCode`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.scanCode.html)，h5端仅支持[微信公众号](https://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421141115)（API以小程序为准），支持 `Promise` 化使用。
 
 **示例代码：**
 
@@ -2861,7 +2861,7 @@ Taro.faceVerifyForPay(params).then(...)
 
 #### Taro.requestPayment(OBJECT)
 
-使用方式同 [`wx.requestPayment`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.requestPayment.html)，支持 `Promise` 化使用。
+使用方式同 [`wx.requestPayment`](https://developers.weixin.qq.com/miniprogram/dev/api/wx.requestPayment.html)，h5端仅支持[微信公众号](https://mp.weixin.qq.com/wiki?t=resource/res_main&id=mp1421141115)（API以小程序为准），支持 `Promise` 化使用。
 
 **示例代码：**
 
@@ -2876,7 +2876,7 @@ Taro.requestPayment(params).then(...)
 | API | 微信小程序 | H5 | React Native |
 | :-: | :-: | :-: | :-: |
 | Taro.faceVerifyForPay | ✔️ |  |  |
-| Taro.requestPayment | ✔️ |  |  |
+| Taro.requestPayment | ✔️ | ✔️ |  |
 
 ### 用户信息
