+++ packages/taro-h5/src/api/open/index.js
@@ -1,0 +1,5 @@
+import { processApis } from '../utils'
+
+export const requestPayment = processApis('chooseWXPay', undefined, undefined, options => {
+  return Object.assign(options, { timestamp: Number.parseInt(options.timeStamp, 10) })
+})
