+++ packages/taro-h5/src/api/unsupportedApi/index.js
@@ -216,7 +216,7 @@ export const checkIsSupportFacialRecognition = temporarilyNotSupport('checkIsSup
 export const startFacialRecognitionVerify = temporarilyNotSupport('startFacialRecognitionVerify')
 export const startFacialRecognitionVerifyAndUploadVideo = temporarilyNotSupport('startFacialRecognitionVerifyAndUploadVideo')
 export const faceVerifyForPay = temporarilyNotSupport('faceVerifyForPay')
-export const requestPayment = temporarilyNotSupport('requestPayment')
+// export const requestPayment = temporarilyNotSupport('requestPayment')
 export const showShareMenu = temporarilyNotSupport('showShareMenu')
 export const hideShareMenu = temporarilyNotSupport('hideShareMenu')
 export const updateShareMenu = temporarilyNotSupport('updateShareMenu')
